refactor(platform-meta): add platform meta getter
