+++ internal/node/image.go
@@ -9,6 +9,7 @@ import (
 
 	cosmodrome "github.com/robolaunch/cosmodrome/pkg/api"
 	"github.com/robolaunch/robot-operator/internal"
+	"github.com/robolaunch/robot-operator/internal/label"
 	robotv1alpha1 "github.com/robolaunch/robot-operator/pkg/api/roboscale.io/v1alpha1"
 	robotv1alpha2 "github.com/robolaunch/robot-operator/pkg/api/roboscale.io/v1alpha2"
 	corev1 "k8s.io/api/core/v1"
@@ -95,8 +96,9 @@ func GetImageForRobot(ctx context.Context, r client.Client, node corev1.Node, ro
 
 	} else {
 
-		platformVersion := GetPlatformVersion(node)
-		imageProps, err := getImagePropsForRobot(ctx, r, platformVersion, getDistroStr(robot.Spec.RobotConfig.Distributions))
+		platformMeta := label.GetPlatformMeta(&node)
+
+		imageProps, err := getImagePropsForRobot(ctx, r, platformMeta.Version, getDistroStr(robot.Spec.RobotConfig.Distributions))
 		if err != nil {
 			return "", err
 		}
@@ -125,8 +127,9 @@ func GetImageForBridge(ctx context.Context, r client.Client, node corev1.Node, r
 	var imageBuilder strings.Builder
 	var tagBuilder strings.Builder
 
-	platformVersion := GetPlatformVersion(node)
-	imageProps, err := getImagePropsForRobot(ctx, r, platformVersion, getDistroStr(robot.Spec.RobotConfig.Distributions))
+	platformMeta := label.GetPlatformMeta(&node)
+
+	imageProps, err := getImagePropsForRobot(ctx, r, platformMeta.Version, getDistroStr(robot.Spec.RobotConfig.Distributions))
 	if err != nil {
 		return "", err
 	}
@@ -151,8 +154,9 @@ func GetBridgeImage(ctx context.Context, r client.Client, node corev1.Node, ros2
 	var imageBuilder strings.Builder
 	var tagBuilder strings.Builder
 
-	platformVersion := GetPlatformVersion(node)
-	imageProps, err := getImagePropsForRobot(ctx, r, platformVersion, getDistroStr([]robotv1alpha1.ROSDistro{ros2Bridge.Spec.Distro}))
+	platformMeta := label.GetPlatformMeta(&node)
+
+	imageProps, err := getImagePropsForRobot(ctx, r, platformMeta.Version, getDistroStr([]robotv1alpha1.ROSDistro{ros2Bridge.Spec.Distro}))
 	if err != nil {
 		return "", err
 	}
@@ -185,8 +189,9 @@ func GetImageForEnvironment(ctx context.Context, r client.Client, node corev1.No
 
 	} else {
 
-		platformVersion := GetPlatformVersion(node)
-		imageProps, err := getImagePropsForEnvironment(ctx, r, platformVersion)
+		platformMeta := label.GetPlatformMeta(&node)
+
+		imageProps, err := getImagePropsForEnvironment(ctx, r, platformMeta.Version)
 		if err != nil {
 			return "", err
 		}
