+++ internal/platform/image.go
@@ -1,0 +1,1 @@
+package platform
